style: setprofile과 myprofileedit 교체
