+++ src/pages/SetProfile/index.jsx
@@ -1,26 +1,37 @@
 import React, { useState, useEffect } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
-import Header from "../../components/Header";
-import Prev from "../../components/Header/Prev";
 import styled from "styled-components";
-import ProfileImg from "../../components/ImageBox";
 import InputBox from "../../components/InputBox/index";
-import axios from "axios";
+import LongBtn from "../../components/Button/LongBtn";
+import ProfileImg from "../../components/ImageBox";
+import { useNavigate, useLocation } from "react-router-dom";
 import userImg from "../../assets/images/user_img_big.svg";
-import MiddleSmallBtn from "./../../components/Button/MiddleSmallBtn/MiddleSmallBtn";
+import axios from "axios";
 
 const FormContainer = styled.section`
-  position: absolute;
-  top: 49px;
   display: flex;
   align-items: center;
   flex-direction: column;
-  /* height: 90vh; */
+  height: 90vh;
   width: 100vw;
+  margin-top: 30px;
   font-family: "Spoqa Han Sans Neo", "sans-serif";
 `;
 
 const PageTitle = styled.h1``;
+
+const ProfileEditTitle = styled.h2`
+  font-weight: 400;
+  font-size: 24px;
+  line-height: 30px;
+  margin-bottom: 1.2rem;
+  color: #000;
+`;
+
+const TitleExplain = styled.p`
+  font-weight: 400;
+  font-size: 14px;
+  color: #767676;
+`;
 
 const ImgWrapper = styled.div`
   width: 110px;
@@ -32,12 +43,7 @@ const InputForm = styled.form`
   margin-bottom: 30px;
 `;
 
-export default function SetProfile() {
-  const [isActive, setIsActive] = useState(false);
-  const [disabled, setDisabled] = useState(false);
-  const location = useLocation();
-  // console.log(location.state);
-
+const SetProfile = () => {
   const [username, setUsername] = useState("");
   const [accountname, setAccountname] = useState("");
   const [itemIntro, setItemIntro] = useState("");
@@ -46,10 +52,12 @@ export default function SetProfile() {
   const [usernameValid, setUsernameValid] = useState(false);
   const [accountnameValid, setAccountnameValid] = useState(false);
   const navigate = useNavigate();
-
+  const location = useLocation();
+  console.log(location);
   const email = location.state.email;
   const password = location.state.password;
   const passed = username && accountname;
+  console.log(email, "비번: ", password);
 
   /* 사용자이름이 바뀔때마다 유효성검사 진행 */
   useEffect(() => {
@@ -78,9 +86,9 @@ export default function SetProfile() {
     }
   }, [accountname]);
 
-  const saveHandler = async () => {
+  const signInHandler = async () => {
     try {
-      await axios.post(`${process.env.REACT_APP_API_KEY}/user`, {
+      const res = await axios.post(`${process.env.REACT_APP_API_KEY}/user`, {
         headers: {
           "Content-type": "application/json",
         },
@@ -102,46 +110,45 @@ export default function SetProfile() {
   };
 
   return (
-    <>
-      <Header>
-        <Prev />
-        <MiddleSmallBtn
-          message="다음"
-          onclick={saveHandler}
-          disabled={passed ? false : true}
+    <FormContainer>
+      <PageTitle className="ir">프로필 설정화면</PageTitle>
+      <ProfileEditTitle>프로필 설정</ProfileEditTitle>
+      <TitleExplain>나중에 언제든지 변경할 수 있습니다.</TitleExplain>
+      <ImgWrapper>
+        <ProfileImg />
+      </ImgWrapper>
+      <InputForm>
+        <InputBox
+          id="username"
+          labelText="사용자 이름"
+          placeholder="2~10자 이내여야 합니다."
+          onChange={(e) => {
+            setUsername(e.target.value);
+          }}
         />
-      </Header>
-      <FormContainer>
-        <PageTitle className="ir">프로필 수정 페이지</PageTitle>
-        <ImgWrapper>
-          <ProfileImg />
-        </ImgWrapper>
-        <InputForm>
-          <InputBox
-            id="username"
-            labelText="사용자 이름"
-            placeholder="2~10자 이내여야 합니다."
-            onChange={(e) => {
-              setUsername(e.target.value);
-            }}
-          />
-          <p className="message">{usernameError}</p>
-          <InputBox
-            id="accountname"
-            labelText="계정 ID"
-            placeholder="영문, 숫자, 특수문자(.),(_)만 사용 가능합니다."
-            onChange={(e) => {
-              setAccountname(e.target.value);
-            }}
-          />
-          <p className="message">{accountnameError}</p>
-          <InputBox
-            id="itemIntro"
-            labelText="소개"
-            placeholder="자신과 판매할 상품에 대해 소개해 주세요!"
-          />
-        </InputForm>
-      </FormContainer>
-    </>
+        <p className="message">{usernameError}</p>
+        <InputBox
+          id="accountname"
+          labelText="계정 ID"
+          placeholder="영문, 숫자, 특수문자(.),(_)만 사용 가능합니다."
+          onChange={(e) => {
+            setAccountname(e.target.value);
+          }}
+        />
+        <p className="message">{accountnameError}</p>
+        <InputBox
+          id="itemIntro"
+          labelText="소개"
+          placeholder="자신과 판매할 상품에 대해 소개해 주세요!"
+        />
+      </InputForm>
+      <LongBtn
+        message="triplus 시작하기"
+        onClick={signInHandler}
+        disabled={passed ? false : true}
+      ></LongBtn>
+    </FormContainer>
   );
-}
+};
+
+export default SetProfile;
